Moved internal functions to header
